Conflicts:
	SztucznaIntCw/SztucznaIntCw/Classes/Diet.cs
	SztucznaIntCw/SztucznaIntCw/Enums/FoodConsumptionTime.cs

+++ SztucznaIntCw/SztucznaIntCw/Classes/NonAbstract/Meal.cs
@@ -5,17 +5,17 @@ namespace SztucznaIntCw.Classes.NonAbstract
 {
     public class Meal : IMeal
     {
-        public int TotalGramsOfProteins { get; set; }
-        public int TotalGramsOfFats { get; set; }
-        public int TotalGramsOfCarbs { get; set; }
+        public decimal TotalGramsOfProteins { get; set; }
+        public decimal TotalGramsOfFats { get; set; }
+        public decimal TotalGramsOfCarbs { get; set; }
 
         //public List<IProduct> ProductList { get;  set; }
-        public Dictionary<IProduct, float> MealProducts { get; set; }
+        public Dictionary<IProduct, decimal> MealProducts { get; set; }
 
         public Meal()
         {
             //ProductList = new List<IProduct>();
-            MealProducts = new Dictionary<IProduct, float>();
+            MealProducts = new Dictionary<IProduct, decimal>();
         }
